requirements.txt
sqlalchemy
python-dotenv
groq
fastapi
uvicorn
httpx
pydantic
requests
python-multipart
jinja2
aiofiles